Removed inline css from Search.jsx
