+++ src/components/Search/Search.css
@@ -4,6 +4,21 @@
   margin-top: 60px;
 }
 
+.search_options {
+  border-style: solid;
+  border-color: gray;
+  border-width: 0.5px;
+  border-radius: 5px;
+  padding: 10px;
+  margin-bottom: 15px;
+}
+
+.search_options h2 {
+  text-align: center;
+  margin: 0;
+  margin-bottom: 10px;
+}
+
 .search_location {
   display: flex;
   flex-direction: row;
@@ -11,19 +26,7 @@
   min-height: 20px;
 }
 
-.search_skills {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 10px;
- }
-
-.search_title {
-  text-align: center;
-  margin: 0;
-}
-
-.search_options {
+.search_options_radio {
   margin-left: 10px;
   display: flex;
   align-items: center;
@@ -65,6 +68,18 @@
   background-color: #f6f5f5;
 }
 
+.result_container {
+  margin-left: 5px;
+}
+
+.result_percent {
+  min-width: 60px;
+}
+
+.clear_button {
+  display: flex;
+  justify-content: flex-end;
+}
 .clear_button button {
   width: 100px;
   height: 40px;
@@ -75,4 +90,29 @@
 
 .clear_button button:hover {
   background-color: #bdbaba;
+}
+
+.result_profile_image img {
+  width: 60px;
+  height: 60px;
+  object-fit: cover;
+  border-radius: 5px;
+}
+
+.result_no_image img {
+  width: 50px;
+  height: 50px;
+  object-fit: cover;
+  border-style: solid;
+  border-color: gray;
+  border-width: 1px;
+  border-radius: 5px;
+}
+
+.result_info {
+  margin-left: 5px;
+}
+
+.result_skills_btns {
+  display: flex;
 }
