+++ src/components/Search/SearchSkills/SearchSkills.css
@@ -1,0 +1,47 @@
+.search_skills_container {
+  border-style: solid;
+  border-color: gray;
+  border-width: 0.5px;
+  border-radius: 5px;
+  padding: 10px;
+  margin-bottom: 15px;
+}
+
+.search_skills_title {
+  text-align: center;
+  margin: 0;
+}
+
+.search_skills {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 10px;
+ }
+
+ .clear_skills_button button {
+  width: 100px;
+  height: 40px;
+  cursor: pointer;
+  border: none;
+  border-radius: 5px;
+}
+
+.clear_skills_button button:hover {
+  background-color: #bdbaba;
+}
+
+.close_icon {
+  font-size: 18px!important;
+  cursor: pointer;
+  margin-left: 4px;
+  border-style: solid;
+  border-color: #f2f1f1;
+  border-width: 0.5px;
+  border-radius: 50px;
+ 
+}
+
+.close_icon:hover {
+    background-color: #b6b5b5;
+}
