+++ src/constants/skills/developerSkills.js
@@ -1,3 +1,3 @@
-const developerSkills = ["3D Modeling Software","ABAP","A/B Testing","Access API","ActionScript","Ada","ADF","ADK","AdMob","Adobe Software","Adobe Air","Adobe CQ","Adobe Flex","Adobe InDesign","ADO.NET","Agavi","Agile","Agile QA","Airbnb","Airtable","Ajax","Akka","Alexa","Algolia","Algorithm","Altcoin","Amazon API","Amazon EC2","Amazon EMR","Amazon Redshift","Amazon S3","Analytics","Android","Android App","Android NDK","Android SDK","Angular","AngularJS","AngularJS","Animation","Ansible","Apache","Apache Airflow","Apache Click","Apache Cordova","Apache Kafka","Apache OFBiz","Apache Pig","Apache Shale","Apache Sling","Apache Struts 2","Apache Tapestry","Apache Tomcat","Apache Wicket","Apex","API","API Development","Apigee","Apollo","Appcelerator","AppExchange","AppFuse","Appian","Appium","Apple","Apple Pay","Apple Watch","App Store","ArcGIS","Arch Linux","ARCore","Arduino","ARKit","ARM","AI","Asana","ASP.NET","ASP.NET Dynamic Data","ASP.NET MVC","Assembly","Asterisk","Augmented Reality","Aurelia","Auth0","Authorize.Net","Autodesk Revit","Automation","AWK","AWS","AWS CLI","AWS Cloud","AWS CloudWatch","AWS Cognito","AWS DevOps","AWS EMR","AWS IAM","AWS IoT","AWS Kinesis","AWS Lambda","AWS Operations","AWS RDS","Azure","B2B","Babel","Backbone.js","Back-end","Bada","Bamboo","Banshee","Basecamp","Bash","Batch","BFC","BigCommerce","BigQuery","Bing Maps","Bitbucket","Bitcoin","Bitcoin Cash","Bitcoin","BizTalk","BlackBerry","Blazor","Blender","Blockchain","BlueBream","Bluetooth","Bokeh","BoonEx Dolphin","Boost","Bootstrap","Bottle","Bower","Box","BPM","Braintree","Brand","Breeze.js","Brew","Brutos Framework","BuddyPress","Business Central","Business Intelligence","C","C#","C++","Caching","CakePHP","Camping","Capistrano","Capybara","Carma","CartoDB","Cassandra","Catalyst","Caveman2","Celery","CFWheels","Chai","Chart.js","Chatbot","ChatGPT","CherryPy","Chromecast","Chrome Extension","CI/CD","CircleCI","Cisco","Clarion","Clojure","Cloud","Cloud Cost Reduction","Cloudflare","Cloud Infrastructure","Cloud Migration","Cloud Security","Clover","Clustering","CMake","CMS","C# / .NET","C++ / .NET","COBOL","Cocoa","CodeDeploy","CodeIgniter","CodeKit","CoffeeScript","Cognos","ColdBox Platform","ColdFusion","ColdSpring","Computer Vision","Confluence","Consul","Contentful","Core Data","Core Java","Corona SDK","Couchbase","CouchDB","CppCMS","CQRS","Crashlytics","CRM","Cross-Platform","Cryptanalysys","Cryptocurrency","Crystal Programming","Crystal Reports","CSLA","CSS","CSS3","CSS / JavaScript","CTOs","CubicWeb","Cucumber","CUDA","cURL","C++ / VBA","Cyber Security","Cython","D","D3.js","Dagger","Dancer","ÐApp","Dapper","Dart","Dashboard","Data Analysis","Database","Database Integration","Database Migration","Databricks","Datadog","Data","Data Integration","Data Management","Data Migration","Data Modeling","Data Modeling Analysis","Data Science","Data Scraping","Data Visualization","Data Warehouse","DB2","Debian","Debugging","Dedicated","Dedicated Android","Dedicated Front-end","Dedicated iPhone","Dedicated Java","Dedicated Magento","Dedicated Mobile App","Dedicated .NET","Dedicated PHP","Dedicated Web","Dedicated WordPress","Deep Learning","DeFi (Decentralized Finance)","Delphi","Demandware","Design","DevExpress","Devise","DevOps","DevSecOps","DirectX","Discord Bot","Distributed Ledger","Distributed Systems","Django","Docker","DocuSign","Dojo","DotNetNuke","DRF Django REST Framework","Dropbox","Dropwizard","Drupal","DynamoDB","eBay API","Eclipse","Eclipse RAP","ECMAScript 6","eCommerce","EDI","Ejabberd","Elasticsearch","Elastic Stack","Electron.js","Elementor","Elgg","Elixir","ELK Stack","Elm","Embedded Software","Embedded Systems","Ember.js","Enfold WordPress","Enterprise","Episerver","Erlang","ERP","ESLint","Espresso","Ethereum","Ethereum Smart Contract","ETL","Evernote","Excel","Excel Macro","Expedia","ExpressionEngine","Express.js","Extended Reality","ExtJS","F#","Fabric","Facebook Ads","Facebook API","Fastlane","Fat-Free","Fault Tolerance","FFmpeg","Financial Software","Fintech","Firebase","Firefox","Firestore","Firmware","Fitbit","Flash","Flask","Flex","Flutter","Flux","Force.com","FormEngine","Fortran","FoxPro","FPGA","FreeBSD","Freelance","Friendster","Front-end","Front-end Web","FTP","FuelPHP","Full-stack","Full Stack Java","Full Stack .NET","Full Stack Node.js","Full Stack PHP","Full Stack Python","Fusebox","Game","Ganache","GatsbyJS","GCC","Gentoo","GIMP","GIS","Git","GitFlow","GitHub","GitOps","GLSL","Gmail","Go","Golang","Google Software","Google AdWords","Google Analytics","Google API","Google App Engine","Google Cardboard","Google Checkout","Google Cloud","Google Cloud Platform","Google Docs","Google Drive","Google Glass","Google Kubernetes Engine (GKE)","Google Latitude","Google Maps","Google Play","Google Sheets","Google Sheets API","Google Sites","Google Storage","Google Web Toolkit","GPT-3","Gradle","Grafana","Grails","GraphQL","Gravity Forms","Grok","Groovy","Groupon","gRPC","Grunt.js","Gulp","Hadoop","HAML","Hamlets","Hapi.js","Happstack","Haproxy","Haskell","Haxe","Hazaar MVC","HDFS","Healthcare Data Science","HealthKit","Helm","Heroku","HFT","Hi5","Hibernate","Highcharts","Hive","HTC Vive","HTML5","Hubspot","HubSpot API","Hugo","Hybrid App","Hybrid Cloud","Hybris","Hyperledger","Hyperledger Fabric","iAd","iBeacon","IBM Watson","iCloud","Identity & Access Management (IAM)","IIS","ImageMagick","Image Processing Software","InfluxDB","Informatica","Infrastructure","Infusionsoft","Instagram API","Integration","Intercom","IoT","Ionic","iOS","iOS / Android","iOS App Development","iPad","iPhone","iPhone App","Iptables","IT","ItsNat","IT Web","Jade","Jamstack","Jasmine","JasperReports","Java","Java EE","JavaFX","Java / Python","JavaScript","Javascript / React","Javascript / WordPress","JavaServer Faces","JBoss","JBoss Seam","JDBC","Jee","Jekyll","Jenkins","Jest","JHipster","Jinja","JIRA","JMeter","JMS","JNI","Joomla","jQuery","JRuby","JSON","JSP","Jspx-bay","Julia","JUnit","Jupyter","JVx","Jython","Kajona","Keras","Keycloak","Kibana","Kindle","Knockout.js","Koa","Kohana","Kotlin","Kubernetes","LabVIEW","LAMP","Laravel","Layar","LDAP","Leaflet","Leaner Style Sheets (LESS)","Leap Motion","LearnDash","Lift","Lightning","LinkedIn API","LINQ","Linux","Linux CentOS","Liquid","Lisp","Litecoin","Lithium","LMS","Load Testing","Lodash","Logstash","Lotus Notes","Lua","Lumen","Mach-II","Machine Learning","MacOS","Magento","MailChimp","Mailgun","Makefile","Mapbox","MapKit","MapQuest","MapReduce","MariaDB","Marionette.js","Markdown","Marketo","Mason","Mathematics","MATLAB","Maven","Maya","Maypole","MEAN Stack","Medical Robotics","Mega","Merb","Mercurial","MERN Stack","MetaMask","Meteor","Microservices","Microsoft","Microsoft Access","Microsoft Dynamics CRM","Microsoft Kinect","Migration","Mixed Reality","Mixpanel","Mobile","Mobile App","Mocha","Mocha.js","Model-Glue","MVVM","MODx","Mojolicious","Moment.js","Moneybookers","MongoDB","Mongoose","Mono","MonoDevelop","MonoRail","Moodle","MPI","MS SQL","MuleSoft","Multi-Cloud","Multithreading","Mustache","MySQL","Nagare","Nagios","NativeScript","Neo4j","NestJS",".NET","NetBeans","NetBSD",".NET Core","NetSuite","Nette Framework","Network Programming","Neural Network","New Relic","Next.js","Nexus","NFT","Nginx","NLP","Node.js","Node.js / Raspberry Pi","nopCommerce","NoSQL","Notion","NPM","NumPy","Nuxt.js","OAuth","Objective-C","Object-oriented Programming","OCaml","Octave","Oculus Rift","Odoo","Offshore","Offshore Android","Offshore App","Offshore iOS","Offshore Java","Offshore Magento","Offshore .NET","Offshore Node.js","Offshore PHP","Offshore Python","Offshore Ruby on Rails","Offshore Web","Offshore Web Design","Offshore WordPress","OneDrive","OpenAI","OpenAI Gym","OpenBSD","OpenCart","OpenCV","OpenERP","OpenGL","OpenLaszlo","OpenRasta","OpenSocial","Open Source","OpenStack","OpenUI5","OpenXava","Optical Character Recognition","Oracle","Oracle ADF","Oracle Forms","Oracle SQL","osCommerce","Outsourced","P2P","Packer","Padrino","Pandas","Pardot","Parse","Pascal","Passport.js","Payment Processing","PayPal","Pebble","Perl","Phoenix","PhoneGap","PHP","phpBB","phpFox","PHPixie","PHP / Laravel","PHP / MySQL","PhpStorm","PHPUnit","PHP Unit Testing","PHP / WordPress","Play","Play Framework","Plone","PL/SQL","Pocket","Poco","Polymer","POS","PostgreSQL","Postman","PowerApps","Power Automate","Power BI","PowerBuilder","PRADO","Preact","PrestaShop","Prezi","Prism","Prisma","Product Architecture","Prolog","Prometheus","Protobuf","Prototype","Protractor","PubNub","Pug","Puppeteer","PureMVC","Pusher","PWA","PyCharm","Pyjs","Pylons","PyObjC","PyQt","Pyramid","Pyspark","Pytest","Python","Python / C#","Pytorch","QA","QA Analysis","QA","Qlik","QlikView","QML","Qt","Quantitative","QuickBase","QuickBooks","R","R3 Corda","RabbitMQ","Rackspace","Ractive.js","Raspberry Pi","Razor","React Hooks","React.js","React Native","React / Node.js","React Web","Realm","Recommendation Systems","Redis","Redmine","Redshift","Redux","Redux Saga","Regex","Release","Remote","Requests","RequireJS","Responsive Web","REST API","RethinkDB","Retrofit","Riak","RIFE","Robotics","ROS","RSpec","RSS","RStudio","RStudio Shiny","Ruby","Ruby on Rails","Rust","RxJava","RxJS","RxSwift","SaaS","Saetta Web Server","Sage","SageMaker","Sails.js","Salesforce","Salesforce1 Mobile App","Salesforce API","Salesforce Commerce Cloud","Salesforce Lightning","SaltStack","SAML","SAS","Sass","SCADA","Scala","Scalatra","Scientific Software","Scikit-Learn","SciPy","SCM","Scrapy","Scrum","SDETs","SecOps","Security","Selenium","Sencha Touch","SendGrid","SendGrid API","Sendmail","Sentry","SEO","Sequelize","Serverless","ServiceNow","SharePoint","Shell","Shell Script","Shopify","Shopify API","Shopify Theme","Sidekiq","SignalR","Silex","Silverlight","SilverStripe","Sinatra","Sinon.JS","Sitefinity","Site Reliability","Site Speed Optimization","Slack","Slim","Smart Contract","Smarty","SMS","Snap","Snowflake","SOA","SOAP","Socket.IO","Software Interface","Solaris","Solidity","SOLIDWORKS","Solr","Sonarqube","SOQL","SourceTree","Spark","SPARQL","Spatial Data Science","Splunk","Spotfire","Spotify","Spree Commerce","Spring","Spring Boot","SpriteKit","SQL","SQL CLR","SQLite","SQL Server","SQL Tuning","Squarespace","SSH","SSIS","SSRS","Startup","Stripe","Stripes","Struts","Stylus","Sublime Text","Subversion (SVN)","SugarCRM","SUN","Svelte","SVG","Swagger","Swift","Sybase","Symbian","Symfony","Sys Administration","System Integration","System Security","Systems","Tableau","TACTIC","Takes","Talend","Teamcity","Tech Freelance","Technical Lead","Technical Writing","Telegram API","Telerik","Tendermint","TensorFlow","Teradata","Terraform","Test Automation","Test-driven Development","ThinWire","Three.js","Thymeleaf","TIBCO","Titanium","Tizen","Tntnet","TokBox","Tornado","Trading Software","Trello","T-SQL","Tumblr","TurboGears","tvOS","Twig","Twilio","Twitter","Twitter Bootstrap","TypeScript","TYPO3","TYPO3 Flow","Uber","Ubuntu","UDP","UI","UIKit","UiPath","UI/UX","UI, UX, and Front-end","Umbraco","UML","Underscore.js","Unified.js","Unit Testing","Unity or Unity3D","Unix","Unreal Engine","Usability","User Acceptance Testing (UAT)","UX","Vaadin","Vagrant","Vanilla JS","Varnishdan","VBA","VB.NET","VBScript","vBulletin","Velocity","Verilog","VHDL","Vibe.d","Vimeo","Vine","Virtualization","Virtual Reality","Visio","Visual Basic","Visualforce","Visual Studio","VMware","VoIP","VoIP Software","Volusion","VRaptor","VSS","vTiger","Vue.js","Vue Storefront","Vuetify.js","Vuforia","Wavemaker","Web","Web2Py","Web3.js","Web Application","Web Architecture","WebAssembly","Webflow","WebGL","WebKit","WebObjects","webOS","Web QA","WebRTC","Web Scraping","Web Services","Website Auditing","WebSphere","WebStorm","Webware","WebWork","WebXR","Weebly","Windows","Windows App","Windows Phone","WinForms","Wireshark","Wix","WooCommerce","WordPress","WordPress Genesis","WordPress Plugin","Wordpress","WordPress Theme","Wowza","WPF","WSDL","Wt","Xamarin","Xamarin.Forms","XAML","Xbox","Xcode","Xero","XHTML","XML","XMPP","XNA","XP","XPath","xQuery","XSD","XSLT","xUnit","Yahoo!","Yahoo! Store","YAML","Yelp API","Yesod","Yii","Yii 2","YouTube API","Zabbix","Zapier","Ze Framework","Zen Cart","Zend","Zendesk","ZK","Zoho","Zoom","Zoom API","Zope 2","Zsh","zTemplates","ZURB Foundation"]
+const developerSkills = ["3D Modeling Software","ABAP","A/B Testing","Access API","ActionScript","Ada","ADF","ADK","AdMob","Adobe Software","Adobe Air","Adobe CQ","Adobe Flex","Adobe InDesign","ADO.NET","Agavi","Agile","Agile QA","Airbnb","Airtable","Ajax","Akka","Alexa","Algolia","Algorithm","Altcoin","Amazon API","Amazon EC2","Amazon EMR","Amazon Redshift","Amazon S3","Analytics","Android","Android App","Android NDK","Android SDK","Angular","AngularJS","AngularJS","Animation","Ansible","Apache","Apache Airflow","Apache Click","Apache Cordova","Apache Kafka","Apache OFBiz","Apache Pig","Apache Shale","Apache Sling","Apache Struts 2","Apache Tapestry","Apache Tomcat","Apache Wicket","Apex","API","API Development","Apigee","Apollo","Appcelerator","AppExchange","AppFuse","Appian","Appium","Apple","Apple Pay","Apple Watch","App Store","ArcGIS","Arch Linux","ARCore","Arduino","ARKit","ARM","AI","Asana","ASP.NET","ASP.NET Dynamic Data","ASP.NET MVC","Assembly","Asterisk","Augmented Reality","Aurelia","Auth0","Authorize.Net","Autodesk Revit","Automation","AWK","AWS","AWS CLI","AWS Cloud","AWS CloudWatch","AWS Cognito","AWS DevOps","AWS EMR","AWS IAM","AWS IoT","AWS Kinesis","AWS Lambda","AWS Operations","AWS RDS","Azure","B2B","Babel","Backbone.js","Back-end","Bada","Bamboo","Banshee","Basecamp","Bash","Batch","BFC","BigCommerce","BigQuery","Bing Maps","Bitbucket","Bitcoin","Bitcoin Cash","Bitcoin","BizTalk","BlackBerry","Blazor","Blender","Blockchain","BlueBream","Bluetooth","Bokeh","BoonEx Dolphin","Boost","Bootstrap","Bottle","Bower","Box","BPM","Braintree","Brand","Breeze.js","Brew","Brutos Framework","BuddyPress","Business Central","Business Intelligence","C","C#","C++","Caching","CakePHP","Camping","Capistrano","Capybara","Carma","CartoDB","Cassandra","Catalyst","Caveman2","Celery","CFWheels","Chai","Chart.js","Chatbot","ChatGPT","CherryPy","Chromecast","Chrome Extension","CI/CD","CircleCI","Cisco","Clarion","Clojure","Cloud","Cloud Cost Reduction","Cloudflare","Cloud Infrastructure","Cloud Migration","Cloud Security","Clover","Clustering","CMake","CMS","C# / .NET","C++ / .NET","COBOL","Cocoa","CodeDeploy","CodeIgniter","CodeKit","CoffeeScript","Cognos","ColdBox Platform","ColdFusion","ColdSpring","Computer Vision","Confluence","Consul","Contentful","Core Data","Core Java","Corona SDK","Couchbase","CouchDB","CppCMS","CQRS","Crashlytics","CRM","Cross-Platform","Cryptanalysys","Cryptocurrency","Crystal Programming","Crystal Reports","CSLA","CSS","CSS3","CSS / JavaScript","CTOs","CubicWeb","Cucumber","CUDA","cURL","C++ / VBA","Cyber Security","Cython","D","D3.js","Dagger","Dancer","ÐApp","Dapper","Dart","Dashboard","Data Analysis","Database","Database Integration","Database Migration","Databricks","Datadog","Data","Data Integration","Data Management","Data Migration","Data Modeling","Data Modeling Analysis","Data Science","Data Scraping","Data Visualization","Data Warehouse","DB2","Debian","Debugging","Dedicated","Dedicated Android","Dedicated Front-end","Dedicated iPhone","Dedicated Java","Dedicated Magento","Dedicated Mobile App","Dedicated .NET","Dedicated PHP","Dedicated Web","Dedicated WordPress","Deep Learning","DeFi (Decentralized Finance)","Delphi","Demandware","Design","DevExpress","Devise","DevOps","DevSecOps","DirectX","Discord Bot","Distributed Ledger","Distributed Systems","Django","Docker","DocuSign","Dojo","DotNetNuke","DRF Django REST Framework","Dropbox","Dropwizard","Drupal","DynamoDB","eBay API","Eclipse","Eclipse RAP","ECMAScript 6","eCommerce","EDI","Ejabberd","Elasticsearch","Elastic Stack","Electron.js","Elementor","Elgg","Elixir","ELK Stack","Elm","Embedded Software","Embedded Systems","Ember.js","Enfold WordPress","Enterprise","Episerver","Erlang","ERP","ESLint","Espresso","Ethereum","Ethereum Smart Contract","ETL","Evernote","Excel","Excel Macro","Expedia","ExpressionEngine","Express.js","Extended Reality","ExtJS","F#","Fabric","Facebook Ads","Facebook API","Fastlane","Fat-Free","Fault Tolerance","FFmpeg","Financial Software","Fintech","Firebase","Firefox","Firestore","Firmware","Fitbit","Flash","Flask","Flex","Flutter","Flux","Force.com","FormEngine","Fortran","FoxPro","FPGA","FreeBSD","Freelance","Friendster","Front-end","Front-end Web","FTP","FuelPHP","Full-stack","Full Stack Java","Full Stack .NET","Full Stack Node.js","Full Stack PHP","Full Stack Python","Fusebox","Game","Ganache","GatsbyJS","GCC","Gentoo","GIMP","GIS","Git","GitFlow","GitHub","GitOps","GLSL","Gmail","Go","Golang","Google Software","Google AdWords","Google Analytics","Google API","Google App Engine","Google Cardboard","Google Checkout","Google Cloud","Google Cloud Platform","Google Docs","Google Drive","Google Glass","Google Kubernetes Engine (GKE)","Google Latitude","Google Maps","Google Play","Google Sheets","Google Sheets API","Google Sites","Google Storage","Google Web Toolkit","GPT-3","Gradle","Grafana","Grails","GraphQL","Gravity Forms","Grok","Groovy","Groupon","gRPC","Grunt.js","Gulp","Hadoop","HAML","Hamlets","Hapi.js","Happstack","Haproxy","Haskell","Haxe","Hazaar MVC","HDFS","Healthcare Data Science","HealthKit","Helm","Heroku","HFT","Hi5","Hibernate","Highcharts","Hive","HTC Vive","HTML5","Hubspot","HubSpot API","Hugo","Hybrid App","Hybrid Cloud","Hybris","Hyperledger","Hyperledger Fabric","iAd","iBeacon","IBM Watson","iCloud","Identity & Access Management (IAM)","IIS","ImageMagick","Image Processing Software","InfluxDB","Informatica","Infrastructure","Infusionsoft","Instagram API","Integration","Intercom","IoT","Ionic","iOS","iOS / Android","iOS App Development","iPad","iPhone","iPhone App","Iptables","IT","ItsNat","IT Web","Jade","Jamstack","Jasmine","JasperReports","Java","Java EE","JavaFX","Java / Python","JavaScript","Javascript / React","Javascript / WordPress","JavaServer Faces","JBoss","JBoss Seam","JDBC","Jee","Jekyll","Jenkins","Jest","JHipster","Jinja","JIRA","JMeter","JMS","JNI","Joomla","jQuery","JRuby","JSON","JSP","Jspx-bay","Julia","JUnit","Jupyter","JVx","Jython","Kajona","Keras","Keycloak","Kibana","Kindle","Knockout.js","Koa","Kohana","Kotlin","Kubernetes","LabVIEW","LAMP","Laravel","Layar","LDAP","Leaflet","Leaner Style Sheets (LESS)","Leap Motion","LearnDash","Lift","Lightning","LinkedIn API","LINQ","Linux","Linux CentOS","Liquid","Lisp","Litecoin","Lithium","LMS","Load Testing","Lodash","Logstash","Lotus Notes","Lua","Lumen","Mach-II","Machine Learning","MacOS","Magento","MailChimp","Mailgun","Makefile","Mapbox","MapKit","MapQuest","MapReduce","MariaDB","Marionette.js","Markdown","Marketo","Mason","Mathematics","MATLAB","Maven","Maya","Maypole","MEAN Stack","Medical Robotics","Mega","Merb","Mercurial","MERN Stack","MetaMask","Meteor","Microservices","Microsoft","Microsoft Access","Microsoft Dynamics CRM","Microsoft Kinect","Migration","Mixed Reality","Mixpanel","Mobile","Mobile App","Mocha","Mocha.js","Model-Glue","MVVM","MODx","Mojolicious","Moment.js","Moneybookers","MongoDB","Mongoose","Mono","MonoDevelop","MonoRail","Moodle","MPI","MS SQL","MuleSoft","Multi-Cloud","Multithreading","Mustache","MySQL","Nagare","Nagios","NativeScript","Neo4j","NestJS",".NET","NetBeans","NetBSD",".NET Core","NetSuite","Nette Framework","Network Programming","Neural Network","New Relic","Next.js","Nexus","NFT","Nginx","NLP","Node.js","Node.js / Raspberry Pi","nopCommerce","NoSQL","Notion","NPM","NumPy","Nuxt.js","OAuth","Objective-C","Object-oriented Programming","OCaml","Octave","Oculus Rift","Odoo","Offshore","Offshore Android","Offshore App","Offshore iOS","Offshore Java","Offshore Magento","Offshore .NET","Offshore Node.js","Offshore PHP","Offshore Python","Offshore Ruby on Rails","Offshore Web","Offshore Web Design","Offshore WordPress","OneDrive","OpenAI","OpenAI Gym","OpenBSD","OpenCart","OpenCV","OpenERP","OpenGL","OpenLaszlo","OpenRasta","OpenSocial","Open Source","OpenStack","OpenUI5","OpenXava","Optical Character Recognition","Oracle","Oracle ADF","Oracle Forms","Oracle SQL","osCommerce","Outsourced","P2P","Packer","Padrino","Pandas","Pardot","Parse","Pascal","Passport.js","Payment Processing","PayPal","Pebble","Perl","Phoenix","PhoneGap","PHP","phpBB","phpFox","PHPixie","PHP / Laravel","PHP / MySQL","PhpStorm","PHPUnit","PHP Unit Testing","PHP / WordPress","Play","Play Framework","Plone","PL/SQL","Pocket","Poco","Polymer","POS","PostgreSQL","Postman","PowerApps","Power Automate","Power BI","PowerBuilder","PRADO","Preact","PrestaShop","Prezi","Prism","Prisma","Product Architecture","Prolog","Prometheus","Protobuf","Prototype","Protractor","PubNub","Pug","Puppeteer","PureMVC","Pusher","PWA","PyCharm","Pyjs","Pylons","PyObjC","PyQt","Pyramid","Pyspark","Pytest","Python","Python / C#","Pytorch","QA","QA Analysis","QA","Qlik","QlikView","QML","Qt","Quantitative","QuickBase","QuickBooks","R","R3 Corda","RabbitMQ","Rackspace","Ractive.js","Raspberry Pi","Razor","React Hooks","React.js","React Native","React Web","Realm","Recommendation Systems","Redis","Redmine","Redshift","Redux","Redux Saga","Regex","Release","Remote","Requests","RequireJS","Responsive Web","REST API","RethinkDB","Retrofit","Riak","RIFE","Robotics","ROS","RSpec","RSS","RStudio","RStudio Shiny","Ruby","Ruby on Rails","Rust","RxJava","RxJS","RxSwift","SaaS","Saetta Web Server","Sage","SageMaker","Sails.js","Salesforce","Salesforce1 Mobile App","Salesforce API","Salesforce Commerce Cloud","Salesforce Lightning","SaltStack","SAML","SAS","Sass","SCADA","Scala","Scalatra","Scientific Software","Scikit-Learn","SciPy","SCM","Scrapy","Scrum","SDETs","SecOps","Security","Selenium","Sencha Touch","SendGrid","SendGrid API","Sendmail","Sentry","SEO","Sequelize","Serverless","ServiceNow","SharePoint","Shell","Shell Script","Shopify","Shopify API","Shopify Theme","Sidekiq","SignalR","Silex","Silverlight","SilverStripe","Sinatra","Sinon.JS","Sitefinity","Site Reliability","Site Speed Optimization","Slack","Slim","Smart Contract","Smarty","SMS","Snap","Snowflake","SOA","SOAP","Socket.IO","Software Interface","Solaris","Solidity","SOLIDWORKS","Solr","Sonarqube","SOQL","SourceTree","Spark","SPARQL","Spatial Data Science","Splunk","Spotfire","Spotify","Spree Commerce","Spring","Spring Boot","SpriteKit","SQL","SQL CLR","SQLite","SQL Server","SQL Tuning","Squarespace","SSH","SSIS","SSRS","Startup","Stripe","Stripes","Struts","Stylus","Sublime Text","Subversion (SVN)","SugarCRM","SUN","Svelte","SVG","Swagger","Swift","Sybase","Symbian","Symfony","Sys Administration","System Integration","System Security","Systems","Tableau","TACTIC","Takes","Talend","Teamcity","Tech Freelance","Technical Lead","Technical Writing","Telegram API","Telerik","Tendermint","TensorFlow","Teradata","Terraform","Test Automation","Test-driven Development","ThinWire","Three.js","Thymeleaf","TIBCO","Titanium","Tizen","Tntnet","TokBox","Tornado","Trading Software","Trello","T-SQL","Tumblr","TurboGears","tvOS","Twig","Twilio","Twitter","Twitter Bootstrap","TypeScript","TYPO3","TYPO3 Flow","Uber","Ubuntu","UDP","UI","UIKit","UiPath","UI/UX","UI, UX, and Front-end","Umbraco","UML","Underscore.js","Unified.js","Unit Testing","Unity or Unity3D","Unix","Unreal Engine","Usability","User Acceptance Testing (UAT)","UX","Vaadin","Vagrant","Vanilla JS","Varnishdan","VBA","VB.NET","VBScript","vBulletin","Velocity","Verilog","VHDL","Vibe.d","Vimeo","Vine","Virtualization","Virtual Reality","Visio","Visual Basic","Visualforce","Visual Studio","VMware","VoIP","VoIP Software","Volusion","VRaptor","VSS","vTiger","Vue.js","Vue Storefront","Vuetify.js","Vuforia","Wavemaker","Web","Web2Py","Web3.js","Web Application","Web Architecture","WebAssembly","Webflow","WebGL","WebKit","WebObjects","webOS","Web QA","WebRTC","Web Scraping","Web Services","Website Auditing","WebSphere","WebStorm","Webware","WebWork","WebXR","Weebly","Windows","Windows App","Windows Phone","WinForms","Wireshark","Wix","WooCommerce","WordPress","WordPress Genesis","WordPress Plugin","Wordpress","WordPress Theme","Wowza","WPF","WSDL","Wt","Xamarin","Xamarin.Forms","XAML","Xbox","Xcode","Xero","XHTML","XML","XMPP","XNA","XP","XPath","xQuery","XSD","XSLT","xUnit","Yahoo!","Yahoo! Store","YAML","Yelp API","Yesod","Yii","Yii 2","YouTube API","Zabbix","Zapier","Ze Framework","Zen Cart","Zend","Zendesk","ZK","Zoho","Zoom","Zoom API","Zope 2","Zsh","zTemplates","ZURB Foundation"]
 
 export { developerSkills }
